by Ernest Shaffer:
This release only works with 0.18+ releases of RackTables.

Improvements to code, bugfixes and features
- Fixed bug where single quotes used in log would give PDO errors.
- Count of logs added to Object logs header.
- Link on object name now takes you directly to that object's Object logs tab.
- URL detection added to notes like RackTables Comments, for clickable links.

The purpose of this plugin is to add the ability to attach notes to objects to track changes, service calls and other information to help support the environment. A new tab shows up under the Objects and object level. At the Objects level it will display all entries for all objects sorted by date. At the object level it will display all logs for that object sorted by date.

Installation instructions are included at the top of the script.


git-svn-id: https://racktables.svn.sourceforge.net/svnroot/racktables@3969