Cosmetic changes for my own use. I've renamed the directory this extention lives in to clarifiy who's it is and what it is. (The previous name 'extentions' implied the dir was part of the existing infrastructure of the project.)
I copied the README and created this markdown version.
I also started to fill out some of the content to show what I did to use this extention.
(The author of this extention deserves lots of praise for the work done!)

Corrected attribution.
I didn't like the way the attribution was not "up top" so I corrected it.