csv_import.php: adopt and relicence as GPLv2
This commit follows 27 ported commits of history of the CSV import
plugin originally developed in the SURFsara fork of racktables-contribs
(at https://github.com/sara-nl/racktables-contribs at the time of this
writing).

Freek Dijkstra of SURFsara in service desk ticket number SD-4444 says
they no longer have a use for this code and it is OK to copy the plugin
to upstream (to make it easier for the community to contribute) and to
relicence under GPLv2 (to make it more consistent with RackTables).